Make the whole Popolo file available to collections and objects
This will make an (algorithmically very inefficient) way of finding
related objects possible.  This is essentially what the Ruby gem this is
based on does.  I would ultimately rather a different implementationa
to achieve this (associating all related objects on first loading the
JSON data) but this will do for the moment.