Define __eq__ and __ne__ for ApproxDate
This includes returning true where you have an ApproxDate precise to the
day on the left hand side of a comparison, and a Python datetime.date on
the right hand side.

This is slightly weird behaviour, since you have the following:

  In [1]: from datetime import date

  In [2]: from approx_dates.models import ApproxDate

  In [4]: approx_date = ApproxDate.from_iso8601('2016-12-25')

  In [5]: datetime_date = date(2016, 12, 25)

  In [6]: approx_date == datetime_date
  Out[6]: True

  In [7]: datetime_date == approx_date
  Out[7]: False

... and I don't think there's an easy way to make that symmetrical.
However, this is nice because it happens that the tests in popolo_data
all have the expected datetime.date on the right hand side, so tests
still pass.  We may want to rethink this later, however.