Make access to related objects more efficient
When looking up some related objects, such as the 'person' property on a
Membership, the code would previously go through every person looking
for one with the right 'id'. This commit changes that so that each
collection also stores a mapping from the objects' ID to the object
itself, so the average time complexity of looking up a person is O(1).

Fixes #3