Don't raise exceptions if there are multiple related values
An example of this is the Person.twitter property; some people have
multiple twitter accounts listed, and if you tried to access the
'twitter' property for one of them, MultipleObjectsReturned would
be raised.

This changes the behaviour of such properties so that if there's
no value, None is returned, and if there is one or more, the first is
returned. (Generally, the "primary" value should be the first, I've
been told by @tmtmtmtm.)

This also makes more sense since @property decorated methods shouldn't
be raising exceptions.

Fixes #6

In a subsequent commit I will add readers for returning all such
values, since there are still plenty of use cases where you might want
all of the known values.